+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 dev
 ---
 
-* -
+* Added a new method to the ``Test`` service: ``echo()``. It echoes back the
+  parameters passed via the `test/echo
+  <https://retdec.com/api/docs/test.html#parameter-passing>`_ service.
 
 0.1 (2015-09-13)
 ----------------